+++ fixtures/main.tf
@@ -6,63 +6,117 @@ provider "aws" {
   s3_force_path_style         = true
   access_key                  = "access_key"
   secret_key                  = "secret_key"
+}
 
-  endpoints {
-    s3 = "http://localhost:4572"
+provider "aws" {
+  alias                       = "account"
+  region                      = "eu-west-2"
+  skip_credentials_validation = true
+  skip_requesting_account_id  = true
+  skip_metadata_api_check     = true
+  s3_force_path_style         = true
+  access_key                  = "access_key"
+  secret_key                  = "secret_key"
+}
+
+provider "aws" {
+  alias                       = "dev"
+  region                      = "eu-west-2"
+  skip_credentials_validation = true
+  skip_requesting_account_id  = true
+  skip_metadata_api_check     = true
+  s3_force_path_style         = true
+  access_key                  = "access_key"
+  secret_key                  = "secret_key"
+
+  assume_role {
+    role_arn = "arn:aws:iam::${module.dev.account_id}:role/OrganizationAccountAccessRole"
   }
 }
+
+provider "aws" {
+  alias                       = "testing"
+  region                      = "eu-west-2"
+  skip_credentials_validation = true
+  skip_requesting_account_id  = true
+  skip_metadata_api_check     = true
+  s3_force_path_style         = true
+  access_key                  = "access_key"
+  secret_key                  = "secret_key"
+
+  assume_role {
+    role_arn = "arn:aws:iam::${module.test.account_id}:role/OrganizationAccountAccessRole"
+  }
+}
+
+provider "aws" {
+  alias                       = "iam"
+  region                      = "eu-west-2"
+  skip_credentials_validation = true
+  skip_requesting_account_id  = true
+  skip_metadata_api_check     = true
+  s3_force_path_style         = true
+  access_key                  = "access_key"
+  secret_key                  = "secret_key"
+
+  assume_role {
+    role_arn = "arn:aws:iam::${module.iam.account_id}:role/OrganizationAccountAccessRole"
+  }
+}
+
 
 module "dev" {
   source = "../"
 
-  account_env          = "dev"
-  account_team         = "cia"
-  config_bucket        = "cia-config-bucket"
-  config_sns_topic_arn = aws_sns_topic.splunk_topic.arn
-  root_account_email   = "aws-registration.ons.000@ons.gov.uk"
-  name                 = "dev"
-  splunk_user_arn      = aws_iam_user.splunk_user.arn
-  dns_subdomain        = "dev"
-  master_zone_id       = "M4ST3RZ0N3ID"
-  iam_account_id       = "012345678910"
-  splunk_logs_sqs_arn  = "arn:aws:sqs:us-east-2:444455556666:queue1"
+  account_env        = "dev"
+  account_team       = "cia"
+  root_account_email = "aws.d.registration.000@ons.gov.uk"
+  name               = "dev"
+  dns_subdomain      = "dev"
+  master_zone_id     = "M4ST3RZ0N3ID"
+  iam_account_id     = module.iam.account_id
+
+  providers = {
+    aws         = aws.dev
+    aws.account = aws.account
+  }
 }
 
 module "test" {
   source = "../"
 
-  account_env          = "test"
-  account_team         = "catd"
-  config_bucket        = "catd-config-bucket"
-  config_sns_topic_arn = aws_sns_topic.splunk_topic.arn
-  root_account_email   = "aws-registration.ons.000@ons.gov.uk"
-  name                 = "test"
-  splunk_user_arn      = aws_iam_user.splunk_user.arn
-  dns_subdomain        = "test"
-  master_zone_id       = "M4ST3RZ0N3ID"
-  iam_account_id       = "012345678910"
+  account_env        = "test"
+  account_team       = "catd"
+  root_account_email = "aws.d.registration.000@ons.gov.uk"
+  name               = "test"
+  dns_subdomain      = "test"
+  master_zone_id     = "M4ST3RZ0N3ID"
+  iam_account_id     = module.iam.account_id
+
+  providers = {
+    aws         = aws.testing
+    aws.account = aws.account
+  }
 }
 
 module "iam" {
   source = "../"
 
-  account_env          = "iam"
-  account_team         = "cia"
-  config_bucket        = "cia-config-bucket"
-  config_sns_topic_arn = aws_sns_topic.splunk_topic.arn
-  root_account_email   = "aws-registration.ons.000@ons.gov.uk"
-  name                 = "iam"
-  splunk_user_arn      = aws_iam_user.splunk_user.arn
-  dns_subdomain        = "iam"
-  master_zone_id       = "M4ST3RZ0N3ID"
-  iam_account_id       = "012345678910"
+  account_env        = "iam"
+  account_team       = "cia"
+  root_account_email = "aws.d.registration.000@ons.gov.uk"
+  name               = "iam"
+  dns_subdomain      = "iam"
+  master_zone_id     = "M4ST3RZ0N3ID"
+  iam_account_id     = module.iam.account_id
+  logging_bucket     = aws_s3_bucket.test_logging_bucket.id
+
+  providers = {
+    aws         = aws.iam
+    aws.account = aws.account
+  }
 }
 
-resource "aws_iam_user" "splunk_user" {
-  name = "splunk-user"
+resource "aws_s3_bucket" "test_logging_bucket" {
+  bucket = "test-logging-bucket"
 }
-
-resource "aws_sns_topic" "splunk_topic" {
-  name = "splunk-topic"
-}
-