+++ sns.tf
@@ -1,0 +1,3 @@
+resource "aws_sns_topic" "config" {
+  name = "${var.name}-config-recorder"
+}